backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sangsoo Jin <constantlyjin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
djangorestframework = "^3.12.2"
django-environ = "^0.4.5"
django-cors-headers = "^3.7.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
